refactor(frontend): merge duplicate carousels into features/game/components
CarouselLocalPlayer1/2 and CarouselSoloPlayer1/2 were identical files.
Kept one pair as CarouselPlayer1/2 in features/game/components where
game UI components belong. Removed features/local-solo/ folder.
