+++ srcs/frontend/src/features/game/components/CarouselPlayer1.tsx
