+++ srcs/frontend/src/features/game/components/CarouselPlayer2.tsx
